+++ src/backend/joanie/core/models/products.py
@@ -39,7 +39,8 @@ from joanie.core.models.courses import (
     Enrollment,
     Organization,
 )
-from joanie.core.utils import contract_definition as contract_definition_utility
+from joanie.core.utils import contract_definition as contract_definition_utility, \
+    payment_schedule
 from joanie.core.utils import issuers, webhooks
 from joanie.core.utils.contract_definition import embed_images_in_context
 from joanie.core.utils.payment_schedule import generate as generate_payment_schedule
@@ -247,6 +248,17 @@ class Product(parler_models.TranslatableModel, BaseModel):
         """
         dates = self.get_equivalent_course_run_dates()
         return CourseRun.compute_state(**dates)
+
+    @property
+    def has_withdrawal_period(self):
+        """
+        Return True if the product has a withdrawal period.
+
+        Read the docstring of core.utils.payment_schedule.has_withdrawal_period method
+        for furthermore information.
+        """
+        dates = self.get_equivalent_course_run_dates()
+        return payment_schedule.has_withdrawal_period(timezone.localdate(), dates.get("start").date())
 
     def clean(self):
         """