+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to
 - Support of payment_schedule for certificate products
 - Add boolean field `has_waived_withdrawal_right` to the `Order` model
 - Add link to back office in Django admin topbar
+- Add `has_withdrawal_period` to `Product` model
 
 ### Changed